+++ src/components/add-review-form/add-review-form.tsx
@@ -1,35 +1,29 @@
 import { useState, ChangeEvent, Fragment, FormEvent } from 'react';
 
 export default function AddReviewForm() {
-  const ratingArray: number[] = [];
-  for (let i = 10; i >= 1; i--) {
-    ratingArray.push(i);
-  }
+  const ratingArray = Array.from({length: 10}, (_, index) => 10 - index);
 
   const [formData, setFormData] = useState({
     rating: '5',
     reviewText: '',
   });
 
-  function handleTextChange(evt: ChangeEvent<HTMLTextAreaElement>) {
-    setFormData({ ...formData, reviewText: evt.target.value });
+  function handleInputChange(evt: ChangeEvent<HTMLTextAreaElement | HTMLInputElement>) {
+    const {name, value} = evt.target;
+    setFormData((prevState) => ({...prevState, [name]: value}));
   }
 
-  function handleRatingChange(evt: ChangeEvent<HTMLInputElement>) {
-    setFormData({ ...formData, rating: evt.target.value });
-  }
-
-  function onSubmitHandler(evt: FormEvent<HTMLFormElement>) {
+  function handleSubmit(evt: FormEvent<HTMLFormElement>) {
     evt.preventDefault();
   }
 
   return (
     <div className="add-review">
-      <form action="#" className="add-review__form" onSubmit={onSubmitHandler}>
+      <form action="#" className="add-review__form" onSubmit= {handleSubmit}>
         <div className="rating">
           <div className="rating__stars">
             {ratingArray.map((rating) => (
-              <Fragment key={crypto.randomUUID()}>
+              <Fragment key={rating}>
                 <input
                   className="rating__input"
                   id={`star-${rating}`}
@@ -37,7 +31,7 @@ export default function AddReviewForm() {
                   name="rating"
                   value={rating}
                   checked={parseInt(formData.rating, 10) === rating}
-                  onChange={handleRatingChange}
+                  onChange={handleInputChange}
                 />
                 <label className="rating__label" htmlFor={`star-${rating}`}>
                   Rating {rating}
@@ -49,11 +43,11 @@ export default function AddReviewForm() {
         <div className="add-review__text">
           <textarea
             className="add-review__textarea"
-            name="review-text"
+            name="reviewText"
             id="review-text"
             placeholder="Review text"
             value={formData.reviewText}
-            onChange={handleTextChange}
+            onChange={handleInputChange}
           />
           <div className="add-review__submit">
             <button className="add-review__btn" type="submit">
