+++ src/types/film-details.ts
@@ -1,0 +1,3 @@
+import { Film } from './film';
+
+export type FilmDetails = Pick<Film, 'director' | 'actors' | 'year' | 'genre' | 'duration'>
