+++ src/types/film-info.ts
@@ -1,0 +1,3 @@
+import {Film} from './film';
+
+export type FilmInfo = Omit<Film, 'previewVideoLink' | 'genre' | 'year' | 'duration' | 'reviews'>
