+++ src/types/film.ts
@@ -1,10 +1,4 @@
-export type Review = {
-  id: string;
-  rating: number;
-  text: string;
-  user: string;
-  date: string;
-}
+import { Review } from './review';
 
 export type Film = {
   id: string;
@@ -23,7 +17,3 @@ export type Film = {
   reviews: Review[];
 };
 
-
-export type FilmDetails = Pick<Film, 'director' | 'actors' | 'year' | 'genre' | 'duration'>
-
-export type FilmInfo = Omit<Film, 'previewVideoLink' | 'genre' | 'year' | 'duration' | 'reviews'>
