+++ src/types/review.ts
@@ -1,0 +1,7 @@
+export type Review = {
+  id: string;
+  rating: number;
+  text: string;
+  user: string;
+  date: string;
+}
